Script to update lookup files for new HODs
